web/app/requirements.txt
CommonMark
dateutils
Django
django-ckeditor
django-crispy-forms
django-datatrans
django-extensions
django-grappelli
django-model-utils
django-modeltranslation
django-pandas
django-rest-swagger
django-reversion
django-tables2
djangorestframework
fuzzywuzzy
gunicorn
httplib2
ipython
Jinja2
numpy
Pillow
psycopg2
PyPDF2
python-dateutil
python-memcached
pytz
PyYAML
qrcode
redis
selenium
setuptools
simplejson
sorl-thumbnail
Sphinx
unicodecsv
Unidecode
XlsxWriter
xlwt
mysqlclient
psycopg2
Wand
ghostscript
